requirements.txt
# Core Django
Django==5.2.4

# Django REST Framework
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1

# API Documentation
drf-spectacular==0.28.0

# CORS
django-cors-headers==4.7.0

# Filtering
django-filter==25.1

# WebSocket Support
channels==4.2.2
channels-redis==4.3.0
daphne==3.0.2

# Redis
redis==6.2.0

# Development tools
django-debug-toolbar==4.2.0
django-extensions==3.2.3

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0

# Code quality
black==23.11.0
flake8==6.1.0
isort==5.12.0

# Production
gunicorn==21.2.0
psycopg2-binary==2.9.9

# Utilities
requests==2.31.0
websockets==15.0.1 